fix:NoData에서 select박스 잘리는 문제해결, type분리 -#44
