+++ src/pages/salary-correction/types/correctionData.ts
@@ -1,0 +1,8 @@
+export interface CorrectionData {
+	month: string;
+	date: string;
+	status: string;
+	history: string | number;
+	correctionType: string;
+	reason: string;
+}
