--- src/pages/salary/types/Graph.ts
@@ -1,4 +1,0 @@
-export interface GraphProps {
-	salaryPercent: number;
-	deductionPercent: number;
-}
