--- src/pages/salary/types/SalaryDetailSection.ts
@@ -1,15 +1,0 @@
-export interface SalaryData {
-	baseSalary: number;
-	nationalPension: number;
-	healthInsurance: number;
-	longTermCareInsurance: number;
-	employmentInsurance: number;
-	incomeTax: number;
-	localIncomeTax: number;
-}
-
-export interface SalaryDetailsSectionProps {
-	salaryData: SalaryData | null;
-	formatNumber: any;
-	totalDeductions: any;
-}
