--- src/pages/salary/types/SalarySection.ts
@@ -1,7 +1,0 @@
-export interface SalarySectionProps {
-	actualPayment: number;
-	isCorrectionModalOpen: boolean;
-	setIsCorrectionModalOpen: (isOpen: boolean) => void;
-	formatNumber: (value: number) => string;
-	selectedDate: any;
-}
