--- src/pages/salary/types/header.ts
@@ -1,7 +1,0 @@
-export interface HeaderProps {
-	selectedDate: Date;
-	setSelectedDate: (date: Date) => void;
-	today: Date;
-	isMonthlySalaryOpen: boolean;
-	setIsMonthlySalaryOpen: (isOpen: boolean) => void;
-}
