+++ src/pages/salary/types/salaryData.ts
