+++ src/pages/salary/types/salaryPage.ts
@@ -1,0 +1,26 @@
+import { SalaryData } from "./salaryData";
+
+export interface GraphProps {
+	salaryPercent: number;
+	deductionPercent: number;
+}
+export interface HeaderProps {
+	selectedDate: Date;
+	setSelectedDate: (date: Date) => void;
+	today: Date;
+	isMonthlySalaryOpen: boolean;
+	setIsMonthlySalaryOpen: (isOpen: boolean) => void;
+}
+export interface SalaryDetailsSectionProps {
+	salaryData: SalaryData | null;
+	formatNumber: any;
+	totalDeductions: any;
+}
+export interface SalarySectionProps {
+	actualPayment: number;
+	isCorrectionModalOpen: boolean;
+	setIsCorrectionModalOpen: (isOpen: boolean) => void;
+	formatNumber: (value: number) => string;
+	selectedDate: any;
+}
+
